requirements.txt
python-dotenv==0.19.2
loguru~=0.6.0
peewee==3.15.3
requests~=2.28.1
urllib3==1.26.12
pyTelegramBotAPI==4.4.0